+++ js/main.js
@@ -1,4 +1,4 @@
-import { closePhotoEditor, initUploadModal, submitForm } from './popup-form.js';
+import { initUploadModal } from './popup-form.js';
 import { renderPictures } from './pictures';
 import { renderBigPhoto } from './popup.js';
 import { getData } from './api.js';
@@ -11,4 +11,3 @@ getData()
 
 
 initUploadModal();
-submitForm(closePhotoEditor);
